requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
httpx==0.28.1
beautifulsoup4==4.13.4
pydantic==2.11.7
rank-bm25==0.2.2
numpy==2.1.1
scikit-learn==1.7.2
pypdf==5.9.0
python-dotenv==1.1.1
pytest==8.4.1
